post_service/requirements.txt
grpcio~=1.71.0
grpcio-tools~=1.71.0
protobuf~=5.29.5
sqlalchemy~=2.0.41
psycopg2-binary~=2.9.10
python-dotenv~=1.1.0
pytest~=8.3.5
kafka-python~=2.2.10